% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_IATA.R
\docType{data}
\name{Data_IATA}
\alias{Data_IATA}
\title{Data_IATA}
\format{An object of class \code{data.frame} with 8965 rows and 6 columns.}
\usage{
Data_IATA
}
\description{
Data_IATA
}
\details{
The dataframe contains IATA and ICAO codes, and geographic information
for 8965 airports around the world. It has been scraped from Wikipedia
https://en.wikipedia.org/wiki/List_of_airports_by_IATA_and_ICAO_code
}
\keyword{datasets}
