% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_iata.R
\name{where_iata}
\alias{where_iata}
\title{IATA localization}
\usage{
where_iata(IATA_code)
}
\arguments{
\item{IATA_code}{a string or vector of strings containing the three letters IATA code(s)}
}
\value{
A dataframe containing IATA code, ICAO codes, Airport name, Location served, Time zone, and DST.

Some of the fields may be `na`.
}
\description{
where_iata answers the question: "Where is located the airport with this IATA code?".
}
\examples{
where_iata("YVR")
where_iata(c("YVR","CHC"))

}
