% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whereport-package.R
\docType{package}
\name{whereport-package}
\alias{whereport}
\alias{whereport-package}
\title{whereport: Retrieve geographical information for airports using IATA codes.}
\description{
Retrieve geographical information for airports using IATA codes.
}
\author{
\strong{Maintainer}: Giulio Valentino Dalla Riva \email{me@gvdallariva.net}

}
\keyword{internal}
