% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data_IATA}
\alias{Data_IATA}
\title{Data_IATA}
\format{A data frame with 8965 rows and 6 variables:
\describe{
  \item{IATA}{IATA code of the airport}
  \item{ICAO}{ICAO code of the airport}
  \item{Airport_name}{Common name of the airport}
  \item{Location_served}{Geographical location for the airport}
  \item{Time}{Time zone of the location served}
  \item{DST}{Months in which the Daylight Saving Time is observed}
}}
\source{
\url{https://en.wikipedia.org/wiki/List_of_airports_by_IATA_and_ICAO_code}
}
\usage{
Data_IATA
}
\description{
Geographical location and Iata, ICAO codes for 8965 international airports
}
\keyword{datasets}
